refactor(stty): remove redundant casts in parse_save_format
- Use parse_hex_u32 return type directly for flag bitfields
- Simplifies code by eliminating unnecessary u32-to-u64 conversions

+++ src/uu/stty/src/stty.rs
@@ -733,10 +733,10 @@ fn parse_save_format(s: &str) -> Result<Termios, Box<dyn UError>> {
         })
     }
 
-    let iflags_bits = parse_hex_u32(parts[0], s)? as u64;
-    let oflags_bits = parse_hex_u32(parts[1], s)? as u64;
-    let cflags_bits = parse_hex_u32(parts[2], s)? as u64;
-    let lflags_bits = parse_hex_u32(parts[3], s)? as u64;
+    let iflags_bits = parse_hex_u32(parts[0], s)?;
+    let oflags_bits = parse_hex_u32(parts[1], s)?;
+    let cflags_bits = parse_hex_u32(parts[2], s)?;
+    let lflags_bits = parse_hex_u32(parts[3], s)?;
 
     // Remaining segments are control_chars.
     let cc_hex = &parts[4..];
